working upGradient and downGradient on linearBoundary and linearBoundaryUScaled grids update the GPL header to LGPL header
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@188 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/grid/storage/hashmap/HashGridStorage.hpp
@@ -13,7 +13,7 @@
 /* sgpp is distributed in the hope that it will be useful,                   */
 /* but WITHOUT ANY WARRANTY; without even the implied warranty of            */
 /* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the             */
-/* GNU General Public License for more details.                              */
+/* GNU Lesser General Public License for more details.                       */
 /*                                                                           */
 /* You should have received a copy of the GNU General Public License         */
 /* along with sgpp; if not, write to the Free Software                       */
@@ -171,7 +171,7 @@ public:
 		return DIM;
 	}
 
-	unsigned int& operator[](index_pointer index)
+	size_t& operator[](index_pointer index)
 	{
 		return map[index];
 	}
